+++ lib/types/user-types.ts
@@ -3,7 +3,3 @@ export interface FormOwner {
   email: string;
   active: boolean;
 }
-
-export enum UserRole {
-  PROGRAM_ADMINISTRATOR = "program_administator",
-}