+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Validation of a JSON Config to check the IDs of elements [#892](https://github.com/cds-snc/platform-forms-client/pull/892)
 - Added login page [#867](https://github.com/cds-snc/platform-forms-client/issues/867)
 - Added login page for temporary token [#900](https://github.com/cds-snc/platform-forms-client/pull/900)
+- Added file attachments to retrieval API [#909](https://github.com/cds-snc/platform-forms-client/pull/909)
+- Logout Page [#847] (https://github.com/cds-snc/platform-forms-client/issues/870)
+- Add Acceptable use page [#873](https://github.com/cds-snc/platform-forms-client/issues/873)
 
 ### Fixed
 
@@ -22,8 +25,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Make GC Branding in Footer configurable [#847](https://github.com/cds-snc/platform-forms-client/pull/847)
-- Logout Page [#847] (https://github.com/cds-snc/platform-forms-client/issues/870)
-- Add Acceptable use page [#873](https://github.com/cds-snc/platform-forms-client/issues/873)
 
 ### Fixed